Move tests into lib/copy/__tests__
