+++ lib/copy/__tests__/fixtures/a-file
