+++ lib/copy/__tests__/fixtures/a-folder/another-file
