+++ lib/copy/__tests__/fixtures/a-folder/another-folder/file3
