+++ lib/copy/__tests__/sync/copy-sync.test.js
@@ -301,4 +301,48 @@ describe('+ copySync()', function () {
       })
     })
   })
+
+  describe('> modification option', function () {
+    var SRC_FIXTURES_DIR = path.join(__dirname, '/../fixtures')
+    var FILES = ['a-file', path.join('a-folder', 'another-file'), path.join('a-folder', 'another-folder', 'file3')]
+
+    describe('> when modified option is turned off', function () {
+      it('should have different timestamps on copy', function (done) {
+        var from = path.join(SRC_FIXTURES_DIR)
+
+        fs.copySync(from, TEST_DIR, {preserveTimestamps: false})
+
+        FILES.forEach(testFile({preserveTimestamps: false}))
+        done()
+      })
+    })
+
+    describe('> when modified option is turned on', function () {
+      it('should have the same timestamps on copy', function (done) {
+        var from = path.join(SRC_FIXTURES_DIR)
+
+        fs.copySync(from, TEST_DIR, {preserveTimestamps: true})
+
+        FILES.forEach(testFile({preserveTimestamps: true}))
+
+        done()
+      })
+    })
+
+    function testFile (options) {
+      return function (file) {
+        var a = path.join(SRC_FIXTURES_DIR, file)
+        var b = path.join(TEST_DIR, file)
+        var fromStat = fs.statSync(a)
+        var toStat = fs.statSync(b)
+        if (options.preserveTimestamps) {
+          assert.deepEqual(toStat.mtime, fromStat.mtime)
+          assert.deepEqual(toStat.atime, fromStat.atime)
+        } else {
+          assert.notEqual(toStat.mtime, fromStat.mtime)
+          assert.notEqual(toStat.atime, fromStat.atime)
+        }
+      }
+    }
+  })
 })
